feat: Command parsing

+++ lib/bot/commander.ex
@@ -1,15 +1,21 @@
 defmodule Lilith.Bot.Commander do
   @moduledoc false
-  @prefix "::"
+  @prefix "%"
 
-  @spec parse(String.t()) :: {String.t(), list(String.t())}
+  @spec parse(String.t()) :: nil | {String.t(), list(String.t())}
   def parse(content) do
-    args = content
-    |> String.slice(String.length(@prefix)..-1)
-    |> String.split(" ")
-    command = args |> List.first()
-    args = args |> List.delete_at(0)
+    if String.starts_with?(content, @prefix) do
+      args =
+        content
+        |> String.slice(String.length(@prefix)..-1)
+        |> String.split(" ")
 
-    {command, args}
+      command = args |> List.first()
+      args = args |> List.delete_at(0)
+
+      {command, args}
+    end
+
+    nil
   end
 end
